Split app.rs monolith into focused submodules
Break the 3,319-line src/app.rs into src/app/ directory module with
five focused files: mod.rs (types, bootstrap, state helpers), input.rs
(keyboard/mouse handling), render.rs (all rendering), sessions.rs
(project/session CRUD), and workers.rs (background threads). No logic
or signature changes — only code movement and pub(crate) visibility.
Update AGENTS.md to document the new module structure and remove the
brittle Key Files list.
